requirements.txt
websockets>=15
PyYAML>=6.0.2